add task-client to rest gateway

+++ goal-service-api/src/main/java/com/benjaminrperry/goalquest/api/goal/Goal.java
@@ -1,9 +1,10 @@
 package com.benjaminrperry.goalquest.api.goal;
 
 public interface Goal {
-    public Integer getId();
+    public Long getId();
     public String getDescription();
     public boolean isCompleted();
+    
 
     public void setCompleted(boolean completed);
 }
